Addressed hash CI build issues.
